# docs/api/signalville-openapi.yaml
openapi: 3.1.0
info:
  title: API SignalVille
  version: 1.1.0
  description: >
    API REST de la plateforme SignalVille, destinée à la gestion des incidents urbains,
    des utilisateurs, des interventions, des notifications et des statistiques.
servers:
  - url: http://localhost:8080/api/v1
    description: Environnement local
...
security:
  - bearerAuth: []
paths:
...
  /reports:
    post:
      tags: [Signalements]
      summary: Créer un signalement
      description: Réservé aux citoyens authentifiés.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateReportMultipart'
      responses:
        '201':
          description: Signalement créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Signalements]
      summary: Lister les signalements visibles par l'utilisateur
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReportStatus'
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPage'
  /reports/{reportId}:
    get:
      tags: [Signalements]
      summary: Consulter un signalement
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Détail du signalement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Signalements]
      summary: Modifier un signalement au statut NOUVEAU
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateReportMultipart'
      responses:
        '200':
          description: Signalement modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '409':
          description: Modification impossible selon le statut courant
  /reports/{reportId}/intervention:
    get:
      tags: [Interventions]
      summary: Consulter l'intervention active d'un signalement
      description: >
        Retourne l'intervention actuellement active (AFFECTEE ou EN_COURS) du
        signalement, si elle existe. Mêmes règles de visibilité que la
        consultation du signalement lui-même.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Intervention active trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionResponse'
        '404':
          description: Signalement introuvable, ou aucune intervention active pour ce signalement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{reportId}/cancel:
    post:
      tags: [Signalements]
      summary: Annuler un signalement au statut NOUVEAU
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Signalement annulé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/{reportId}/reject:
    post:
      tags: [Interventions]
      summary: Rejeter un signalement
      description: Réservé au superviseur, uniquement depuis le statut NOUVEAU.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Signalement rejeté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/{reportId}/assign:
    post:
      tags: [Interventions]
      summary: Affecter un signalement à un agent
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignReportRequest'
      responses:
        '201':
          description: Intervention créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionResponse'
  /reports/{reportId}/reassign:
    post:
      tags: [Interventions]
      summary: Réaffecter un incident
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReassignReportRequest'
      responses:
        '201':
          description: Nouvelle intervention créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionResponse'
  /reports/{reportId}/start:
    post:
      tags: [Interventions]
      summary: Démarrer l'intervention
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Intervention démarrée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionResponse'
  /reports/{reportId}/resolve:
    post:
      tags: [Interventions]
      summary: Déclarer l'incident résolu
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ResolveReportMultipart'
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Incident déclaré résolu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionResponse'
  /reports/{reportId}/close:
    post:
      tags: [Interventions]
      summary: Clôturer un incident résolu
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                publicComment:
                  type: string
      responses:
        '200':
          description: Incident clôturé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /reports/{reportId}/reopen:
    post:
      tags: [Interventions]
      summary: Rouvrir un incident
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Incident rouvert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Lister ses notifications
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Notifications paginées
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPage'
  /notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Marquer une notification comme lue
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification marquée comme lue
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RoleName:
      type: string
      enum: [CITOYEN, AGENT, SUPERVISEUR, ADMINISTRATEUR]
    AccountStatus:
      type: string
      enum: [ACTIF, SUSPENDU, DESACTIVE]
    ReportStatus:
      type: string
      enum: [NOUVEAU, AFFECTE, EN_COURS, RESOLU, CLOTURE, REOUVERT, REJETE, ANNULE]
    InterventionStatus:
      type: string
      enum: [AFFECTEE, EN_COURS, RESOLUE, REAFFECTEE, INTERROMPUE]
    Priority:
      type: string
      enum: [BASSE, MOYENNE, HAUTE, CRITIQUE]
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: '#/components/schemas/RoleName'
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: [string, 'null']
          format: date-time
...
    CategoryRequest:
      type: object
      required: [name, defaultPriority, targetDelayHours]
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        defaultPriority:
          $ref: '#/components/schemas/Priority'
        targetDelayHours:
          type: integer
          minimum: 1
        active:
          type: boolean
          default: true
    CategoryResponse:
      allOf:
        - $ref: '#/components/schemas/CategoryRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
    CreateReportMultipart:
      type: object
      required: [title, description, categoryId, latitude, longitude, address, photos]
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 30
          maxLength: 1000
        categoryId:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        district:
          type: string
        municipality:
          type: string
        photos:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
            format: binary
    UpdateReportMultipart:
      allOf:
        - $ref: '#/components/schemas/CreateReportMultipart'
    ResolveReportMultipart:
      type: object
      required: [resolutionComment, proofs]
      properties:
        resolutionComment:
          type: string
          minLength: 10
        proofs:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
            format: binary
    AssignReportRequest:
      type: object
      required: [agentId]
      properties:
        agentId:
          type: string
          format: uuid
        instruction:
          type: string
    ReassignReportRequest:
      type: object
      required: [newAgentId, reason]
      properties:
        newAgentId:
          type: string
          format: uuid
        reason:
          type: string
    ReportResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          example: SV-2026-000001
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ReportStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        category:
          $ref: '#/components/schemas/CategoryResponse'
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        district:
          type: string
        municipality:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReportDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ReportResponse'
        - type: object
          properties:
            citizen:
              $ref: '#/components/schemas/UserResponse'
            activeIntervention:
              anyOf:
                - $ref: '#/components/schemas/InterventionResponse'
                - type: 'null'
            history:
              type: array
              items:
                $ref: '#/components/schemas/HistoryResponse'
    ReportPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReportResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    PhotoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        description:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
    InterventionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        agent:
          $ref: '#/components/schemas/UserResponse'
        status:
          $ref: '#/components/schemas/InterventionStatus'
        assignedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        resolvedAt:
          type: [string, 'null']
          format: date-time
        instruction:
          type: string
        resolutionComment:
          type: string
        proofs:
          type: array
          items:
            $ref: '#/components/schemas/PhotoResponse'
...
    HistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        previousStatus:
          anyOf:
            - $ref: '#/components/schemas/ReportStatus'
            - type: 'null'
        newStatus:
          $ref: '#/components/schemas/ReportStatus'
        comment:
          type: string
        actor:
          $ref: '#/components/schemas/UserResponse'
        changedAt:
          type: string
          format: date-time
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        link:
          type: string
        createdAt:
          type: string
          format: date-time
    NotificationPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer

    # -------------------------------------------------------------------
    # Tableaux de bord
    #
    # Un schema typé par rôle, au lieu d'un Map<String, Number> générique.
    # Motif : les écrans superviseur et agent doivent afficher des LISTES
    # de signalements actionnables (à vérifier, critiques, en cours), pas
    # seulement des compteurs. Un dictionnaire de nombres ne peut pas les
    # porter, et le front perdrait tout typage TypeScript.
    #
    # Chaque dashboard expose des compteurs + les listes courtes dont son
    # écran a réellement besoin, afin d'éviter un appel réseau par bloc.
    # -------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string